requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.7.4
pydantic-settings==2.1.0
sqlalchemy==2.0.29
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
python-dateutil==2.8.2
requests==2.31.0
httpx==0.27.2
mcp[cli]==1.1.2
pytest==8.2.0

